Доп проверка в hard delete

+++ src/PetHome.Application/Features/Write/PetManegment/HardDelete/HardDeletePetUseCase.cs
@@ -61,14 +61,16 @@ public class HardDeletePetUseCase
             await _volunteerRepository.Update(volunteer, ct);
             await _unitOfWork.SaveChages(ct);
             transaction.Commit();
-
-            List<MinioFileName> minioFileNames = pet.Medias
-                .Select(f => MinioFileName.Create(f.FileName).Value)
-                .ToList();
-            string bucketName = pet.Medias.Select(f => f.BucketName).First();
-            MinioFilesInfoDto minioFileInfoDto = new MinioFilesInfoDto(bucketName, minioFileNames);
-            await _filesProvider.DeleteFile(minioFileInfoDto, ct);
-
+          
+            if (pet.Medias.Count > 0)
+            {
+                List<MinioFileName> minioFileNames = pet.Medias
+                    .Select(f => MinioFileName.Create(f.FileName).Value)
+                    .ToList();
+                string bucketName = pet.Medias.Select(f => f.BucketName).First();
+                MinioFilesInfoDto minioFileInfoDto = new MinioFilesInfoDto(bucketName, minioFileNames);
+                await _filesProvider.DeleteFile(minioFileInfoDto, ct);
+            }
             string message = $"Питомец = {command.PetId} успешно hard deleted!";
             _logger.LogInformation(message);
             return Result.Success<ErrorList>();
